+++ app/assets/javascripts/profiles.coffee
