+++ app/assets/stylesheets/profiles.scss
