--- app/helpers/profiels_helper.rb
@@ -1,2 +1,0 @@
-module ProfielsHelper
-end
