+++ app/views/profiles/new.html.erb
@@ -1,0 +1,5 @@
+<%= form_for(@profile, url: create_profile_path(current_user.id)) do |f| %>
+    <%= f.label :content %>
+    <%= f.text_field :content %>
+    <%= f.submit "submit" %>
+<% end %>
