+++ app/views/shared/_user_info.html.erb
@@ -1,4 +1,4 @@
 <%= link_to gravatar_for(current_user, size: 50), current_user %>
 <h1><%= current_user.name %></h1>
-<span><%= link_to "view my profile", current_user %></span>
+<span><%= link_to "view my profiles", current_user %></span>
 <span><%= pluralize(current_user.microposts.count, "micropost") %></span>
