+++ app/views/users/show.html.erb
@@ -10,10 +10,17 @@
     <section class="stats">
       <%= render 'shared/stats' %>
     </section>
+    <h3 style="text-align: left">プロフィール</h3>
+    <div class="profile">
     <% unless @profile.nil? %>
-        <%= @profile.content %>
+        <p><%= @profile.content %></p>
     <% end %>
-    <%= link_to "プロフィール編集", "#" %>
+    </div>
+    <% if @profile.nil? %>
+      <%= link_to "プロフィール編集", new_profile_path %>
+    <% else %>
+      <%= link_to "プロフィール編集", edit_profile_path %>
+    <% end %>
   </aside>
 
   <div class="col-md-8">
