+++ config/routes.rb
@@ -1,9 +1,11 @@
 Rails.application.routes.draw do
-  get 'profiels/create'
+  post 'profiles/create/:id', to: 'profiles#create', as: 'create_profile'
 
-  get 'profiels/destroy'
+  post 'profiles/update/:id', to: 'profiles#update', as: 'update_profile'
 
-  get 'profiels/new'
+  get 'profiels/new', to: 'profiles#new', as: 'new_profile'
+
+  get 'profiles/edit', to: 'profiles#edit', as: 'edit_profile'
 
   root                'static_pages#home'
   get    'help'    => 'static_pages#help'
